@@ -22,109 +23,107 @@
     "logdown": "3.3.1",
     "minimist": "1.2.8",
     "open-graph": "0.2.6",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-redux": "8.1.3",
     "redux": "4.2.1",
     "redux-logger": "3.0.6",
     "redux-thunk": "2.4.2",
-    "uuidjs": "4.2.12"
+    "uuid": "9.0.1"
   },
   "description": "The most secure collaboration platform.",
   "devDependencies": {
-    "@babel/core": "7.23.0",
-    "@babel/helper-string-parser": "7.21.5",
+    "@babel/core": "7.26.8",
+    "@babel/helper-string-parser": "7.25.9",
     "@babel/plugin-proposal-class-properties": "7.18.6",
     "@babel/plugin-proposal-optional-chaining": "7.21.0",
-    "@babel/preset-env": "7.22.20",
-    "@babel/preset-react": "7.18.6",
-    "@babel/preset-typescript": "7.23.0",
-    "@babel/register": "7.22.15",
-    "@electron/fuses": "1.6.1",
-    "@types/adm-zip": "0.5.2",
-    "@types/amplify": "1.1.26",
-    "@types/auto-launch": "5.0.1",
-    "@types/content-type": "1.1.6",
-    "@types/eslint": "^8.44.3",
-    "@types/fs-extra": "9.0.11",
-    "@types/hapi__joi": "^17.1.11",
-    "@types/is-ci": "3.0.1",
-    "@types/jest": "^29.5.5",
-    "@types/lodash": "4.14.199",
-    "@types/minimist": "1.2.3",
-    "@types/mocha": "10.0.2",
-    "@types/node": "18.18.4",
-    "@types/open-graph": "0.2.3",
-    "@types/platform": "1.3.4",
+    "@babel/preset-env": "7.26.8",
+    "@babel/preset-react": "7.26.3",
+    "@babel/preset-typescript": "7.26.0",
+    "@babel/register": "7.25.9",
+    "@electron/fuses": "1.8.0",
+    "@electron/osx-sign": "1.3.2",
+    "@types/adm-zip": "0.5.7",
+    "@types/amplify": "1.1.28",
+    "@types/auto-launch": "5.0.5",
+    "@types/content-type": "1.1.8",
+    "@types/eslint": "^8.56.7",
+    "@types/fs-extra": "11.0.4",
+    "@types/hapi__joi": "^17.1.15",
+    "@types/is-ci": "3.0.4",
+    "@types/jest": "^29.5.14",
+    "@types/lodash": "4.17.15",
+    "@types/minimist": "1.2.5",
+    "@types/mocha": "10.0.10",
+    "@types/node": "18.19.75",
+    "@types/open-graph": "0.2.5",
+    "@types/platform": "1.3.6",
     "@types/prettier": "^2.7.3",
-    "@types/react": "^18.2.25",
-    "@types/react-dom": "^18.2.11",
-    "@types/redux-logger": "^3.0.10",
-    "@types/sinon": "10.0.19",
-    "@types/sort-json": "2.0.1",
-    "@typescript-eslint/eslint-plugin": "6.7.4",
-    "@typescript-eslint/parser": "6.7.4",
-    "@wireapp/copy-config": "1.3.6",
-    "@wireapp/eslint-config": "2.1.2",
-    "@wireapp/prettier-config": "0.6.3",
-    "adm-zip": "0.5.10",
-    "aws-sdk": "2.1472.0",
+    "@types/react": "^18.3.13",
+    "@types/react-dom": "^18.3.1",
+    "@types/redux-logger": "^3.0.12",
+    "@types/sinon": "17.0.3",
+    "@types/sort-json": "2.0.3",
+    "@types/uuid": "9.0.8",
+    "@typescript-eslint/eslint-plugin": "7.17.0",
+    "@typescript-eslint/parser": "7.18.0",
+    "@wireapp/copy-config": "2.3.0",
+    "@wireapp/eslint-config": "3.0.21",
+    "@wireapp/prettier-config": "0.6.4",
+    "adm-zip": "0.5.16",
+    "aws-sdk": "2.1692.0",
     "babel-core": "7.0.0-bridge.0",
     "babel-eslint": "10.1.0",
     "babel-jest": "29.7.0",
-    "babel-loader": "9.1.2",
+    "babel-loader": "9.2.1",
     "babel-plugin-istanbul": "6.1.1",
-    "commander": "10.0.1",
-    "core-js": "3.33.0",
+    "commander": "12.1.0",
+    "core-js": "3.40.0",
     "cross-env": "7.0.3",
-    "cspell": "6.31.1",
-    "css-loader": "6.8.1",
-    "dotenv": "16.3.1",
-    "electron": "26.3.0",
-    "electron-builder": "24.6.4",
-    "electron-mocha": "12.1.0",
-    "electron-osx-sign": "0.6.0",
-    "electron-packager": "16.0.0",
+    "css-loader": "7.1.2",
+    "dotenv": "16.4.7",
+    "electron": "34.1.1",
+    "electron-builder": "24.13.3",
+    "electron-mocha": "12.3.1",
+    "electron-packager": "17.1.2",
     "electron-winstaller": "4.0.1",
-    "eslint": "8.51.0",
-    "eslint-config-prettier": "8.10.0",
+    "eslint": "8.57.0",
+    "eslint-config-prettier": "9.1.0",
     "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-better-styled-components": "1.1.2",
     "eslint-plugin-header": "3.1.1",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-jasmine": "4.1.3",
-    "eslint-plugin-jest": "27.4.2",
-    "eslint-plugin-jest-dom": "4.0.3",
-    "eslint-plugin-jsdoc": "39.6.7",
-    "eslint-plugin-jsx-a11y": "6.7.1",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-jasmine": "4.2.2",
+    "eslint-plugin-jest": "28.11.0",
+    "eslint-plugin-jest-dom": "5.5.0",
+    "eslint-plugin-jsdoc": "48.11.0",
+    "eslint-plugin-jsx-a11y": "6.10.2",
     "eslint-plugin-no-unsanitized": "4.0.2",
     "eslint-plugin-prettier": "4.2.1",
-    "eslint-plugin-react": "7.33.2",
-    "eslint-plugin-react-hooks": "4.6.0",
-    "eslint-plugin-simple-import-sort": "8.0.0",
-    "eslint-plugin-sort-keys-fix": "1.1.2",
-    "eslint-plugin-testing-library": "5.10.3",
-    "eslint-plugin-typescript-sort-keys": "2.3.0",
-    "eslint-plugin-unused-imports": "2.0.0",
-    "form-data": "4.0.0",
+    "eslint-plugin-react": "7.37.4",
+    "eslint-plugin-react-hooks": "4.6.2",
+    "eslint-plugin-simple-import-sort": "12.1.1",
+    "eslint-plugin-testing-library": "6.5.0",
+    "eslint-plugin-unused-imports": "3.2.0",
+    "form-data": "4.0.1",
     "husky": "4.3.8",
     "is-ci": "3.0.1",
     "jest": "29.7.0",
     "jest-environment-jsdom": "29.7.0",
-    "lint-staged": "14.0.1",
-    "mocha": "10.2.0",
-    "nock": "13.3.3",
+    "lint-staged": "15.4.3",
+    "mocha": "10.8.2",
+    "nock": "13.5.6",
     "nyc": "15.1.0",
     "prettier": "2.8.8",
-    "rimraf": "5.0.5",
-    "sinon": "16.1.0",
+    "rimraf": "5.0.8",
+    "sinon": "17.0.2",
     "sort-json": "2.0.1",
-    "style-loader": "3.3.3",
-    "ts-node": "10.9.1",
-    "typescript": "4.8.4",
-    "webpack": "5.88.2",
-    "webpack-cli": "5.0.2"
+    "style-loader": "4.0.0",
+    "ts-node": "10.9.2",
+    "typescript": "5.7.3",
+    "webpack": "5.97.1",
+    "webpack-cli": "5.1.4"
   },
   "homepage": "https://wire.com",
   "husky": {
