@@ -24,8 +24,8 @@ import {settings} from '../settings/ConfigurationPersistence';
 import {SettingsType} from '../settings/SettingsType';
 
 const argv = minimist(process.argv.slice(1));
-const customWebappUrl: string | undefined = argv[config.ARGUMENT.ENV];
-const isProdEnvironment = !!customWebappUrl;
+const webappUrlSetting = settings.restore<string | undefined>(SettingsType.CUSTOM_WEBAPP_URL);
+const customWebappUrl: string | undefined = argv[config.ARGUMENT.ENV] || webappUrlSetting;
 
 export enum ServerType {
   PRODUCTION = 'PRODUCTION',
@@ -42,11 +42,6 @@ interface AvailableEnvironment {
 
 let currentEnvironment = settings.restore<ServerType | undefined>(SettingsType.ENV);
 
-const URL_WEBSITE = {
-  PRODUCTION: config.websiteUrl,
-  STAGING: 'https://wire-website-staging.zinfra.io',
-};
-
 const webappEnvironments = {
   [ServerType.PRODUCTION]: {name: 'Production', server: ServerType.PRODUCTION, url: 'https://app.wire.com'},
   [ServerType.BETA]: {name: 'Beta', server: ServerType.BETA, url: 'https://wire-webapp-staging.wire.com'},
@@ -55,6 +50,7 @@ const webappEnvironments = {
 
 export const app = {
   ENV: config.environment,
+  DESKTOP_VERSION: config.version,
   IS_DEVELOPMENT: config.environment !== 'production',
   IS_PRODUCTION: config.environment === 'production',
   UPDATE_URL_WIN: config.updateUrl,
@@ -130,7 +126,7 @@ export function getAvailebleEnvironments(): AvailableEnvironment[] {
 export const web = {
   getWebappUrl,
   getWebsiteUrl: (path: string = ''): string => {
-    const baseUrl = isProdEnvironment ? URL_WEBSITE.PRODUCTION : URL_WEBSITE.STAGING;
+    const baseUrl = config.websiteUrl;
     return `${baseUrl}${path}`;
   },
 };
