@@ -43,6 +43,10 @@ export declare global {
   };
   var environment: typeof EnvironmentUtil;
   var openGraphAsync: (url: string) => Promise<OpenGraphResult>;
+  var desktopAppConfig: {
+    version: string;
+    supportsCallingPopoutWindow?: boolean;
+  };
   /* eslint-enable no-var */
 
   interface Window {
@@ -52,7 +56,7 @@ export declare global {
     locStrings: i18nStrings;
     locStringsDefault: i18nStrings;
     sendBadgeCount(count: number, ignoreFlash: boolean): void;
-    sendConversationJoinToHost(accountId: string, code: string, key: string): void;
+    sendConversationJoinToHost(accountId: string, code: string, key: string, domain?: string): void;
     sendDeleteAccount(accountId: string, sessionID?: string): Promise<void>;
     sendLogoutAccount(accountId: string): Promise<void>;
     submitDeepLink(url: string): void;
@@ -69,6 +73,7 @@ export declare global {
       util: {
         Environment: {
           version(showWrapperVersion: boolean): string;
+          avsVersion?: () => string;
         };
       };
     };
