@@ -17,7 +17,7 @@
  *
  */
 
-import {BrowserWindow, screen, shell} from 'electron';
+import {BaseWindow, BrowserWindow, screen, shell} from 'electron';
 
 import * as path from 'path';
 import {URL} from 'url';
@@ -78,3 +78,85 @@ export const openExternal = async (url: string, httpsOnly: boolean = false): Pro
     logger.error(error);
   }
 };
+
+const isBrowserWindow = (baseWindow: unknown): baseWindow is BrowserWindow => {
+  return baseWindow instanceof Object && 'webContents' in baseWindow;
+};
+
+export const sendToWebContents = (baseWindow: BaseWindow | undefined, channel: string, ...args: any[]) => {
+  if (isBrowserWindow(baseWindow)) {
+    try {
+      baseWindow.webContents.send(channel, ...args);
+    } catch (error) {
+      logger.error('Failed to send event to webContents', error);
+    }
+  } else {
+    logger.error("This action's target is not an instance of BrowserWindow.");
+  }
+};
+
+export const getNewWindowOptions = ({
+  parent,
+  title = '',
+  width,
+  height,
+  resizable = false,
+  fullscreenable = false,
+  maximizable = false,
+  alwaysOnTop = true,
+  minimizable = false,
+  autoHideMenuBar = false,
+}: {
+  title?: string;
+  parent?: Electron.BrowserWindow;
+  width: number;
+  height: number;
+  resizable?: boolean;
+  fullscreenable?: boolean;
+  maximizable?: boolean;
+  alwaysOnTop?: boolean;
+  minimizable?: boolean;
+  autoHideMenuBar?: boolean;
+}): Electron.BrowserWindowConstructorOptions => ({
+  alwaysOnTop,
+  width,
+  height,
+  backgroundColor: '#FFFFFF',
+  fullscreen: false,
+  fullscreenable,
+  maximizable,
+  minimizable,
+  modal: false,
+  movable: true,
+  parent,
+  resizable,
+  minHeight: height,
+  minWidth: width,
+  title: title,
+  titleBarStyle: 'default',
+  useContentSize: true,
+  autoHideMenuBar,
+  webPreferences: {
+    allowRunningInsecureContent: false,
+    backgroundThrottling: false,
+    contextIsolation: true,
+    devTools: false,
+    disableBlinkFeatures: '',
+    experimentalFeatures: false,
+    images: true,
+    javascript: true,
+    nodeIntegration: false,
+    nodeIntegrationInWorker: false,
+    offscreen: false,
+    partition: '',
+    plugins: false,
+    preload: '',
+    sandbox: true,
+    scrollBounce: true,
+    spellcheck: false,
+    textAreasAreResizable: false,
+    webSecurity: true,
+    webgl: false,
+    webviewTag: false,
+  },
+});
