@@ -30,14 +30,14 @@ node('built-in') {
 
   stage('Checkout & Clean') {
     git branch: "${GIT_BRANCH}", url: 'https://github.com/wireapp/wire-desktop.git'
-    sh returnStatus: true, script: 'rm -rf *.pkg *.zip ./wrap/dist/'
+    sh returnStatus: true, script: 'rm -rf *.pkg *.zip ./wrap/dist/ ./node_modules/'
   }
 
   def projectName = env.WRAPPER_BUILD.tokenize('#')[0]
   def version = env.WRAPPER_BUILD.tokenize('#')[1]
   echo('version: ' + version)
   def buildNumber = version.tokenize('.')[2]
-  def NODE = tool name: 'node-v16.17.1', type: 'nodejs'
+  def NODE = tool name: 'node-v20.10.0', type: 'nodejs'
   env.DRY_RUN = params.DRY_RUN ? "--dry-run" : ""
 
   stage('Get build artifacts') {
